Change send_ssl_data to use size_t instead of uint16_t - Commit a299ddc

+++ src/ssl_client.cpp
@@ -353,7 +353,7 @@ int data_to_read(sslclient_context *ssl_client)
 }
 
 
-int send_ssl_data(sslclient_context *ssl_client, const uint8_t *data, uint16_t len)
+int send_ssl_data(sslclient_context *ssl_client, const uint8_t *data, size_t len)
 {
     log_d("Writing SSL (%d bytes)...", len);  //for low level debug
     int ret = -1;
